By making the match_id a simple int we lose the automatic value as well
The reason to make the match_id an int is that in essence we want to be
able to specify "winner of match 40 vs. winner of match 46" where the
match_id is specific to a tournament. As we do not yet have tournaments
we assume a single tournament, but we do need to hardcode which match is
which to make referring to them reliable.

Note that depending on the UI to add matches (CSV vs. web UI) this may
make less sense, but even so most sources assign stable match ids to the
matches of the tournament, no reason for us to act in a different way