Simplify everything by using a view
At the moment diesel doesn't really support views properly, because they
don't have a primary key. This also means it doesn't export the views
into print-schema, which means we add it to the patch method. This means
we no longer require the sql_query function and can compose queries with
the view info.