Allow setting favourites
Show them in dashboard, plus provide a link to set/update them.

Unfortunately it seems that the on conflict approach doesn't work well,
possibly because there are so many constraints on the table.

This means that if you switch one country with another it will fail,
because we first detect the duplicate country. One option is to set all
countries to null before we loop over it.

Remark: the order when fetching the current selection (for the edit
form) is rather important, because if you do not do that then country
switches will happen at random.

Also unfortunately we can't place a span within an option tag, this
means I might need to resort to some kind of advanced JS trickery to let
the England flag look good.