Two fixes for scoring
On the one hand I said that points would be given for correctly
predicting 90/120/penalties, so give a point for that.

And on the other hand some points were kept from people because they
entered predictions using a broken interface, thereby losing points. As
a general rule assume they didn't really think about it and meant to say
90 minutes.